added new class

+++ src/javaCodeForTester/NumberCount.java
@@ -1,4 +1,4 @@
-package javaCodeForTester;
+	package javaCodeForTester;
 
 import java.util.Scanner;
 
@@ -9,9 +9,9 @@ public class NumberCount {
 		Scanner sc = new Scanner(System.in);
 		System.out.println("Enter the Number");
 		int num = sc.nextInt();
-		int count = 0;
-		while (num > 0) {
-			num = num / 10;
+		long count=0;
+		while(num>0) {
+			num=num/10;
 			count++;
 		}
 		System.out.println(count);
